Rewrote readUnprocessedTrips, helper functions - moved functions to timelineHelper - updated momentJS code to luxon
